Часть 4.

+++ src/main/java/ru/skypro/homework/repository/AdRepository.java
@@ -7,5 +7,5 @@ import java.util.List;
 
 public interface AdRepository extends JpaRepository<Ad, Long> {
 
-    List<Ad> findByAuthorId(Integer userId);
+    List<Ad> findByAuthorId(Long userId);
 }
